# src/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Facebook Events API
  version: 1.0.0
servers:
  - url: http://localhost:3000
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      required: [id, email, firstName, lastName]
      properties:
        id:
          type: string
        email:
          type: string
          format: email
        firstName:
          type: string
        lastName:
          type: string
...
    Error:
      type: object
      required: [message]
      properties:
        message:
          type: string
        details:
          type: array
          items:
            type: object
            properties:
              path:
                type: string
              message:
                type: string
    GroupMember:
      type: object
      required: [user, role]
      properties:
        user:
          $ref: '#/components/schemas/User'
        role:
          type: string
          enum: [member, admin, superadmin]
    Group:
      type: object
      required:
        - id
        - name
        - description
        - icon
        - coverPhoto
        - visibility
        - membersCanPost
        - membersCanCreateEvents
        - role
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        icon:
          type: string
          format: uri
          nullable: true
        coverPhoto:
          type: string
          format: uri
          nullable: true
        visibility:
          type: string
          enum: [public, private, secret]
        membersCanPost:
          type: boolean
        membersCanCreateEvents:
          type: boolean
        role:
          type: string
          nullable: true
          enum: [member, admin, superadmin]
        members:
          type: array
          items:
            $ref: '#/components/schemas/GroupMember'
...
    Event:
      type: object
      required:
        - id
        - name
        - description
        - startDate
        - endDate
        - location
        - coverPhoto
        - visibility
        - group
        - createdBy
        - organizers
        - participants
        - shoppingListEnabled
        - carpoolingEnabled
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        location:
          type: string
        coverPhoto:
          type: string
          format: uri
          nullable: true
        visibility:
          type: string
          enum: [public, private]
        group:
          type: string
          nullable: true
        createdBy:
          $ref: '#/components/schemas/User'
        organizers:
          type: array
          items:
            $ref: '#/components/schemas/User'
        participants:
          type: array
          items:
            $ref: '#/components/schemas/User'
        shoppingListEnabled:
          type: boolean
        carpoolingEnabled:
          type: boolean
    EventWrite:
      type: object
      additionalProperties: false
      required: [name, startDate, endDate, location, visibility]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 80
        description:
          type: string
          maxLength: 2000
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        location:
          type: string
          minLength: 1
          maxLength: 200
        coverPhoto:
          type: string
          format: uri
        visibility:
          type: string
          enum: [public, private]
        shoppingListEnabled:
          type: boolean
        carpoolingEnabled:
          type: boolean
...
paths:
...
  /api/groups:
    post:
      summary: Create a group
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [name, visibility]
              properties:
                name:
                  type: string
                  minLength: 1
                  maxLength: 80
                description:
                  type: string
                  maxLength: 2000
                icon:
                  type: string
                  format: uri
                coverPhoto:
                  type: string
                  format: uri
                visibility:
                  type: string
                  enum: [public, private, secret]
                membersCanPost:
                  type: boolean
                membersCanCreateEvents:
                  type: boolean
      responses:
        '201':
          description: Group created. The creator is the superadmin, so members are included.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Group'
        '400':
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Missing or invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    get:
      summary: List visible groups
      description: Public and private groups, plus secret groups the caller belongs to. Each entry is a profile with the caller's role, without the member list.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Visible groups
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Group'
        '401':
          description: Missing or invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/events:
    post:
      summary: Create an event
      description: The caller becomes the creator, an organizer, and a participant.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventWrite'
      responses:
        '201':
          description: Event created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '400':
          description: Validation failed, or startDate is not before endDate
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Missing or invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    get:
      summary: List visible events
      description: Public events, plus private events the caller organizes, attends, created, or can administer as a group admin or superadmin. Sorted by start date.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Visible events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Event'
        '401':
          description: Missing or invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...